fix(frontend): router redirects (#58)
This fixes router redirects breaking the back button. Browser history is now also set to memory history in testing mode, which enables full testing of the router by driving the history instead of the route provider.

The router was modified to use computed values instead of reactions to determine the components, and use autorun only for side effects that dispatch changes to external modules (i.e. history)