Generic RPC error handlers (#73)
Introduce `RPCErrorHandlers` to define custom error handling.
Allows users to define custom error handlers which will handle RPC call errors.
They are registered in a stack-like manner, meaning the handlers which are most recently added will be run first. The mechanism iterates over all error handlers in that order and terminates at the first handlers that is able to handle the error. If no handler is able to handle the error, there are 2 fallback handlers, a CodedError handler, which handles CodedError, and a general 500 error handler which makes sure the error is propagated with status code 500.

Changes made:
- Update `fail()` method in `RPCDispatcher`
  - **Removes** the custom test for joi `if ((e as any).isJoi) {`
  - adds custom error handling above coded error handling
- Existing test that verified custom handling is updated with the new procedure:
  - `app.getSingleton(RPCErrorHandlers).addErrorHandler((e: Error) => {`

Maintenance:
- relaxed bluebird dependency to `"bluebird": "^3.5.1"` (it was `"bluebird": "3.5.1"`)
- removed unused dependency `@types/body-parser`
- moved `@types/content-disposition` to `devDependencies`